Fixed an odd corner case
I ran into a case where the subscription listener was called
after the unsubscribe.  This resulted in setState being called
on unmounted components.  So I introduced an extra flag to
avoid this.  But this is pretty complicated.  A simpler solution
would be to have this handled inside redux which is why I
commented on this issue: https://github.com/rackt/redux/issues/461