Cleaned out and new reactors
I cleaned out the old react and browser stuff by putting it into seperate
modules.  This way, you only pull in those modules and their dependencies
and environmental assumptions when you need to.

I also changed the definition of reactors so that the reactor gets
access to the previous and next state.  This is going to allow much
smarter reactors and also the ability to very cleanly implement
onEnter and onLeave reactors for routing.